+++ examples/kyb-app/src/components/organisms/KYBView/views/CompanyInformationView/const.ts
@@ -1,0 +1,2 @@
+export const COUNTRIES_WITH_STATES = ['ae', 'ca', 'us'];
+export const AUTOFILLABLE_FIELDS = ['companyName', 'companyType', 'vat', 'establishmentDate'];