+++ examples/kyb-app/src/components/organisms/KYBView/views/CompanyInformationView/types.ts
@@ -1,9 +1,9 @@
 export interface CompanyInformationContext {
-  registrationNumber: string;
-  country: string;
-  state: string;
   companyName: string;
-  vat: string;
+  companyCountry: string;
   companyType: string;
+  registrationNumber: string;
+  state: string;
+  vat: string;
   registrationDate: string;
 }